move gpt-2 scripts to models\gpt2
