+++ onnxruntime/python/tools/transformers/__init__.py
@@ -1,4 +1,8 @@
-import os
-import sys
+#-------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation.  All rights reserved.
+# Licensed under the MIT License.
+#--------------------------------------------------------------------------
 
-sys.path.append(os.path.dirname(__file__))
+from . import (
+    models,
+)
