+++ onnxruntime/python/tools/transformers/models/__init__.py
@@ -1,8 +1,10 @@
 #-------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
+# Copyright (c) Microsoft Corporation.  All rights reserved.
 # Licensed under the MIT License.
-#-------------------------------------------------------------------------
-import os
-import sys
+#--------------------------------------------------------------------------
 
-sys.path.append(os.path.dirname(__file__))
+from . import (
+    gpt2,
+    longformer,
+    t5
+)
