+++ onnxruntime/python/tools/transformers/models/gpt2/gpt2_tester.py
@@ -4,16 +4,18 @@
 # license information.
 # --------------------------------------------------------------------------
 # This script helps evaluation of GPT-2 model.
-import os
 import logging
 import torch
-import random
 import numpy
-import time
 import timeit
 import math
 import statistics
 from gpt2_helper import Gpt2Helper, Gpt2Inputs
+
+import sys
+import os
+sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..'))
+
 from benchmark_helper import Precision
 
 logger = logging.getLogger(__name__)
