+++ onnxruntime/python/tools/transformers/models/longformer/__init__.py
@@ -1,10 +1,10 @@
-# -------------------------------------------------------------------------
+#-------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation.  All rights reserved.
-# Licensed under the MIT License.  See License.txt in the project root for
-# license information.
-# --------------------------------------------------------------------------
+# Licensed under the MIT License.
+#--------------------------------------------------------------------------
 
-import os
-import sys
 
-sys.path.append(os.path.dirname(__file__))
+from .longformer_helper import (
+    LongformerInputs,
+    LongformerHelper,
+)
