+++ onnxruntime/python/tools/transformers/models/t5/__init__.py
@@ -1,8 +1,25 @@
 #-------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
+# Copyright (c) Microsoft Corporation.  All rights reserved.
 # Licensed under the MIT License.
-#-------------------------------------------------------------------------
-import os
-import sys
+#--------------------------------------------------------------------------
 
-sys.path.append(os.path.dirname(__file__))
+from .past_helper import (
+    PastKeyValuesHelper,
+)
+
+from .t5_helper import (
+    PRETRAINED_T5_MODELS,
+    T5Helper,
+)
+
+from .t5_encoder import (
+    T5EncoderHelper,
+)
+
+from .t5_decoder import (
+    T5DecoderHelper,
+)
+
+from .t5_encoder_decoder_init import (
+    T5EncoderDecoderInitHelper,
+)
