+++ onnxruntime/test/python/transformers/parity_utilities.py
@@ -10,13 +10,13 @@ import numpy
 import torch
 
 
-def find_transformers_source():
-    source_dir = os.path.join(os.path.dirname(__file__), '..', '..', '..', 'python', 'tools', 'transformers')
+def find_transformers_source(sub_dir_paths=[]):
+    source_dir = os.path.join(os.path.dirname(__file__), '..', '..', '..', 'python', 'tools', 'transformers', *sub_dir_paths)
     if (os.path.exists(source_dir)):
         if source_dir not in sys.path:
             sys.path.append(source_dir)
         return True
-    return False
+    return False    
 
 
 def create_inputs(batch_size=1, sequence_length=1, hidden_size=768, float16=False, device=torch.device('cuda')):
