+++ onnxruntime/test/python/transformers/test_gpt2.py
@@ -13,10 +13,11 @@ import coloredlogs
 import pytest
 
 from parity_utilities import find_transformers_source
-if find_transformers_source():
+
+if find_transformers_source(sub_dir_paths=['models', 'gpt2']):
     from benchmark_gpt2 import parse_arguments, main
 else:
-    from onnxruntime.transformers.benchmark_gpt2 import parse_arguments, main
+    from onnxruntime.transformers.models.gpt2.benchmark_gpt2 import parse_arguments, main
 
 
 class TestGpt2(unittest.TestCase):
