+++ setup.py
@@ -319,8 +319,9 @@ packages = [
     'onnxruntime.quantization.operators',
     'onnxruntime.quantization.CalTableFlatBuffers',
     'onnxruntime.transformers',
-    'onnxruntime.transformers.models.t5',
+    'onnxruntime.transformers.models.gpt2',
     'onnxruntime.transformers.models.longformer',
+    'onnxruntime.transformers.models.t5',    
 ]
 
 requirements_file = "requirements.txt"
